+++ src/index.js
@@ -1,4 +1,6 @@
+import loadList from './modules/load-list.js';
 import './style.css';
+import { delFromLocalStorage, changeLocalStorage, resetIndex } from './modules/local-storage.js';
 
 const form = document.querySelector('.input-form');
 const ul = document.querySelector('.todo-list');
@@ -11,44 +13,6 @@ const saveToLocalStorage = (task, completed, index) => {
 
   taskArr.push(todo);
   localStorage.setItem('todos', JSON.stringify(taskArr));
-};
-
-const delFromLocalStorage = (index) => {
-  let taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-
-  taskArr = taskArr.filter((item) => {
-    if (item.index !== index) {
-      return true;
-    }
-    return false;
-  });
-
-  localStorage.setItem('todos', JSON.stringify(taskArr));
-};
-
-const changeLocalStorage = (task, status, index) => {
-  let taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-  taskArr = taskArr.map((item) => {
-    if (item.index === index) {
-      item.task = task;
-      item.completed = status;
-    }
-    return item;
-  });
-  localStorage.setItem('todos', JSON.stringify(taskArr));
-};
-
-// reset index
-const resetIndex = () => {
-  const taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-  const arr = [];
-
-  taskArr.forEach((obj) => {
-    const newObj = { ...obj, index: (arr.length + 1).toString() };
-    arr.push(newObj);
-  });
-  localStorage.setItem('todos', JSON.stringify(arr));
-  window.location.reload();
 };
 
 const addTodo = (e) => {
@@ -172,95 +136,7 @@ clearBtn.addEventListener('click', () => {
 });
 
 window.addEventListener('DOMContentLoaded', () => {
-  const taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-  if (taskArr.length > 0) {
-    taskArr.forEach((todo) => {
-      const li = document.createElement('li');
-      li.className = 'li-todo';
-      const attr = document.createAttribute('data-index');
-      attr.value = todo.index;
-      li.setAttributeNode(attr);
-      li.innerHTML = `<div class="check">
-    <input type="checkbox" class="checkbox">
-    <div class="parent"><p class="para">${todo.task}</p></div>
-</div>
-<div>
-    <i class="fa fa-ellipsis-v" aria-hidden="true"></i>
-    <i class="fa fa-trash" aria-hidden="true"></i>
-</div>`;
-      ul.appendChild(li);
-      const checkboxes = li.querySelectorAll('.checkbox');
-      checkboxes.forEach((checkbox) => {
-        checkbox.addEventListener('change', (e) => {
-          if (checkbox.checked === true) {
-            const todo = e.target.nextElementSibling.childNodes[0];
-            todo.style.textDecoration = 'line-through';
-            const todoLi = e.target.parentElement.parentElement;
-            const { index } = todoLi.dataset;
-            const taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-            taskArr.filter((todo) => {
-              if (todo.index === index) {
-                todo.completed = true;
-              }
-              return false;
-            });
-            localStorage.setItem('todos', JSON.stringify(taskArr));
-          } else {
-            const todo = e.target.nextElementSibling.childNodes[0];
-            todo.style.textDecoration = 'none';
-            const todoLi = e.target.parentElement.parentElement;
-            const { index } = todoLi.dataset;
-            const taskArr = localStorage.getItem('todos') ? JSON.parse(localStorage.getItem('todos')) : [];
-            taskArr.filter((todo) => {
-              if (todo.index === index) {
-                todo.completed = false;
-              }
-              return false;
-            });
-            localStorage.setItem('todos', JSON.stringify(taskArr));
-          }
-        });
-      });
-
-      const optionBtn = li.querySelector('.fa-ellipsis-v');
-      const trash = li.querySelector('.fa-trash');
-      optionBtn.addEventListener('click', (e) => {
-        optionBtn.classList.add('hide');
-        trash.classList.add('show');
-        const edit = e.target.parentElement.previousElementSibling.lastElementChild.childNodes[0];
-        const editInput = document.createElement('input');
-        editInput.type = 'text';
-        editInput.className = 'changed-task';
-        li.classList.add('edit-task');
-        editInput.value = edit.textContent;
-        const editIndex = e.target.parentElement.parentElement.dataset.index;
-        const parentDiv = e.target.parentElement.previousElementSibling.lastElementChild;
-        parentDiv.removeChild(edit);
-        parentDiv.appendChild(editInput);
-        const completed = false;
-        editInput.addEventListener('keypress', (e) => {
-          if (e.key === 'Enter') {
-            edit.textContent = editInput.value;
-
-            parentDiv.appendChild(edit);
-            parentDiv.removeChild(editInput);
-            li.classList.remove('edit-task');
-            optionBtn.classList.remove('hide');
-            trash.classList.remove('show');
-          }
-          changeLocalStorage(edit.textContent, completed, editIndex);
-        });
-      });
-      trash.addEventListener('click', (e) => {
-        const task = e.target.parentElement.parentElement;
-        const { index } = task.dataset;
-        ul.removeChild(li);
-        delFromLocalStorage(index);
-        resetIndex();
-        window.location.reload();
-      });
-    });
-  }
+  loadList();
 });
 
 function createNewListItem(task) {
